# api-spec.yaml
openapi: 3.0.0
info:
  title: Japanese Learning App API
  version: 1.0.0
  description: |
    Backend API for a Japanese language learning platform.
    Serves kana, kanji, vocabulary, sentences, and components.
    Authenticated users can override kanji mnemonics.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.yourdomain.com
    description: Production

# ─────────────────────────────────────────────────────────────
# AUTH ARCHITECTURE
# ─────────────────────────────────────────────────────────────
#
# Two-token system:
#
#   Access token  — short-lived JWT (15 min). Returned in response
#                   body on login/register. Client stores in JS memory
#                   only (never localStorage, never a cookie). Sent as
#                   Authorization: Bearer header on every protected request.
#                   Server validates cryptographically — no DB hit.
#
#   Refresh token — long-lived JWT (30 days). Stored in DB so it can
#                   be revoked. Set by server as an HttpOnly cookie
#                   (JavaScript cannot read it). Browser/native cookie
#                   handler sends it automatically on /auth/refresh and
#                   /auth/logout. Never appears in any response body.
#
# Proactive refresh (client-side):
#   Before every request, the client decodes the access token JWT and
#   checks the exp claim. If expiry is < 5 minutes away, it calls
...
#
# App startup:
...
#
...
#
...
#
...
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Short-lived access token (15 min). Stored in JS memory only.
        Obtained from login/register response body or refresh response body.
        Send as Authorization: Bearer <token> header.

  # ───────────────────────────────────────────────────────────
  # SHARED PARAMETERS
  # ───────────────────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of results per page
    JlptParam:
      name: jlpt
      in: query
      schema:
        type: string
        enum: [N5, N4, N3, N2, N1]
      description: >
        Optional. Filter to exactly this JLPT level. Mutually exclusive with
        jlpt_max — providing both is a 422.
    JlptMaxParam:
      name: jlpt_max
      in: query
      schema:
        type: string
        enum: [N5, N4, N3, N2, N1]
      description: >
        Optional. Filter to this JLPT level or easier (e.g. N3 includes N3,
        N4, and N5; NULL-jlpt items are excluded). Mutually exclusive with
        jlpt — providing both is a 422.


  # ───────────────────────────────────────────────────────────
  # SCHEMAS
  # ───────────────────────────────────────────────────────────
  schemas:
...
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: not_found
        message:
          type: string
          example: Resource not found
    PaginatedKana:
      type: object
      required: [data, total, page, page_size]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Kana'
        total:
          type: integer
          example: 208
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
    PaginatedKanji:
      type: object
      required: [data, total, page, page_size]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Kanji'
        total:
          type: integer
          example: 2200
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
    PaginatedVocab:
      type: object
      required: [data, total, page, page_size]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Vocab'
        total:
          type: integer
          example: 10000
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
    PaginatedSentence:
      type: object
      required: [data, total, page, page_size]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Sentence'
        total:
          type: integer
          example: 50000
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
    PaginatedComponent:
      type: object
      required: [data, total, page, page_size]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Component'
        total:
          type: integer
          example: 253
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20

    # ── Auth ────────────────────────────────────────────────
...
    Kana:
      type: object
      required: [character, romaji, type, category]
      properties:
        character:
          type: string
          example: き
          description: The kana character — also the primary key
        romaji:
          type: string
          example: ki
        type:
          type: string
          enum: [hiragana, katakana]
          example: hiragana
        row:
          type: string
          nullable: true
          example: k
          description: >
            Chart row (consonant group). Derived at seed time from romaji + category.
            Null for yōon characters (きゃ etc.) and ん/ン.
        col:
          type: string
          nullable: true
          example: i
          description: >
            Chart column (vowel). Derived at seed time from the last character of romaji.
            Null for yōon characters and ん/ン.
        category:
          type: string
          enum: [base, dakuten, handakuten, yoon]
          example: base
        katakana_equivalent:
          type: string
          nullable: true
          example: キ
          description: For hiragana entries — the katakana version
        hiragana_equivalent:
          type: string
          nullable: true
          example: き
          description: For katakana entries — the hiragana version
        mnemonic:
          type: string
          nullable: true
          example: Looks like a key
        audio_url:
          type: string
          nullable: true
          format: uri
          example: https://r2.yourdomain.com/audio/kana/ki.mp3
        stroke_order_svg_url:
          type: string
          nullable: true
          format: uri
          example: https://r2.yourdomain.com/kanjivg/kana/き.svg

    # ── Component (Radical) ─────────────────────────────────
    Component:
      type: object
      required: [id, character, keyword, meaning, stroke_count]
      properties:
        id:
          type: integer
          example: 1
        character:
          type: string
          example: 木
        keyword:
          type: string
          example: tree
          description: Short English label used in mnemonic generation
        meaning:
          type: string
          example: Tree or wood
        stroke_count:
          type: integer
          example: 4

    # ── Kanji ───────────────────────────────────────────────
    Kanji:
      type: object
      required: [character, unicode_hex, meanings, stroke_count]
      properties:
        character:
          type: string
          example: 新
          description: The kanji character — also the primary key
        unicode_hex:
          type: string
          example: 065b0
          description: Unicode codepoint hex — guaranteed present for all JLPT kanji
        meanings:
          type: array
          items:
            type: string
          example: ["new", "neo-"]
          description: English meanings — guaranteed present for all JLPT kanji
        onyomi:
          type: array
          nullable: true
          items:
            type: string
          example: ["シン"]
          description: On'yomi readings — nullable, some kanji have only kun'yomi
        kunyomi:
          type: array
          nullable: true
          items:
            type: string
          example: ["あたら.しい", "あら.た"]
          description: Kun'yomi readings — nullable, some kanji have only on'yomi
        nanori:
          type: array
          nullable: true
          items:
            type: string
          example: ["あらた"]
          description: Name readings — nullable, only present for kanji used in names
        jlpt:
          type: string
          nullable: true
          enum: [N5, N4, N3, N2, N1]
          example: N2
          description: Nullable at insert time — overlaid from Tanos after initial KANJIDIC2 import
        grade:
          type: integer
          nullable: true
          example: 2
          description: Japanese school grade (1–6 elementary, 8 secondary) — nullable, not all JLPT kanji have a school grade
        stroke_count:
          type: integer
          example: 13
          description: Stroke count — guaranteed present for all JLPT kanji
        frequency:
          type: integer
          nullable: true
          example: 56
          description: Frequency rank out of 2500 most-used kanji
        classical_radical_number:
          type: integer
          nullable: true
          example: 69
          description: Kangxi radical number from KANJIDIC2
        stroke_order_svg_url:
          type: string
          nullable: true
          format: uri
          example: https://r2.yourdomain.com/kanjivg/065b0.svg
        mnemonic:
          type: string
          nullable: true
          example: >
            A person stands next to a tree and swings an axe —
            they're clearing land to build something NEW.
          description: LLM-generated default mnemonic
        user_mnemonic:
          type: string
          nullable: true
          example: My new axe stands against the tree outside.
          description: >
            The authenticated user's personal mnemonic override.
            Null means the user has not written one — show mnemonic instead.
            Only present on GET /kanji/{character} (requires auth).
        components:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Component'
          description: >
            Visual components this kanji is made of.
            Only included on GET /kanji/{character} (detail view).
            Not included in list responses for performance.
        sentences:
          type: array
          nullable: true
          maxItems: 10
          items:
            $ref: '#/components/schemas/Sentence'
          description: >
            Up to 10 example sentences, ordered by sentence ID (ascending).
            Only included on GET /kanji/{character}. Not included in list responses.
        vocab_words:
          type: array
          nullable: true
          maxItems: 20
          items:
            $ref: '#/components/schemas/KanjiVocabEntry'
          description: >
            Up to 20 vocabulary words that contain this kanji, ordered by JLPT level
            (N5 first, NULL last) then by is_common (common words first), then by
            vocab id as a stable tie-breaker. Grouped by reading_type on the frontend.
            Only included on GET /kanji/{character}. Not included in list responses.
    KanjiVocabEntry:
      type: object
      required: [id, word, reading, meanings]
      properties:
        id:
          type: string
          description: JMdict entry sequence number
          examples: ["1578850"]
        word:
          type: string
          examples: ["食べる"]
        reading:
          type: string
          examples: ["たべる"]
        romaji:
          type: string
          nullable: true
          examples: ["taberu"]
        meanings:
          type: array
          items:
            $ref: '#/components/schemas/VocabMeaning'
        jlpt:
          type: string
          nullable: true
          enum: [N5, N4, N3, N2, N1]
          examples: ["N5"]
        is_common:
          type: boolean
          nullable: true
          description: True if tagged ichi1/news1/spec1 in JMdict
        reading_type:
          type: string
          nullable: true
          enum: [on, kun]
          description: >
            Whether this vocab word uses the on'yomi or kun'yomi reading of the kanji.
            "on" = Chinese-origin compound reading (e.g. 食事 uses ショク).
            "kun" = Native Japanese reading (e.g. 食べる uses た).
            null = reading could not be resolved (irregular or single-char reading).
...
    VocabMeaning:
      type: object
      properties:
        pos:
          type: array
          items:
            type: string
          example: ["v1"]
          description: Part of speech tags from JMdict
        definitions:
          type: array
          items:
            type: string
          example: ["to eat", "to live on"]
    FuriganaSegment:
      type: object
      required: [ruby]
      properties:
        ruby:
          type: string
          example: 食
        rt:
          type: string
          nullable: true
          example: た
          description: Reading for this segment. Null for kana-only segments.
    Vocab:
      type: object
      required: [id, word, reading, meanings]
      properties:
        id:
          type: string
          example: "1578850"
          description: JMdict entry sequence number
        word:
          type: string
          example: 食べる
        reading:
          type: string
          example: たべる
        romaji:
          type: string
          nullable: true
          example: taberu
        meanings:
          type: array
          items:
            $ref: '#/components/schemas/VocabMeaning'
        furigana:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/FuriganaSegment'
        jlpt:
          type: string
          nullable: true
          enum: [N5, N4, N3, N2, N1]
          example: N5
        is_common:
          type: boolean
          nullable: true
          example: true
          description: True if tagged ichi1/news1/spec1 in JMdict
        sentences:
          type: array
          nullable: true
          maxItems: 10
          items:
            $ref: '#/components/schemas/Sentence'
          description: >
            Up to 10 example sentences, ordered by sentence ID (ascending).
            Only included on GET /vocab/{id}. Not included in list responses.

    # ── Files (R2 uploads) ───────────────────────────────────
    #
    # Upload flow:
    #   1. Client calls POST /files/presign with filename/mime_type/size_bytes.
    #      Server validates size against MAX_FILE_SIZE_BYTES and returns a
    #      presigned R2 PUT URL + the object_key the client must use.
    #   2. Client PUTs the file bytes directly to upload_url (bytes never
    #      pass through the API).
    #   3. Client calls POST /files/confirm with the same object_key plus
    #      filename/mime_type/size_bytes. Server writes the `files` metadata
    #      row and returns it.
...
    Sentence:
      type: object
      required: [id, japanese, english]
      properties:
        id:
          type: integer
          example: 1234
          description: Tatoeba sentence ID
        japanese:
          type: string
          example: 私は毎日学校に行きます。
        english:
          type: string
          example: I go to school every day.

    # ── Reading comprehension ───────────────────────────────
    ReadingQuestion:
      type: object
      required: [id, passage_id, question_order, question_text, options, correct_answer, explanation]
      properties:
        id:
          type: integer
          example: 1
        passage_id:
          type: integer
          example: 1
        question_order:
          type: integer
          example: 1
        question_text:
          type: string
          example: "田中さんはなぜ電話をしましたか。"
        options:
          type: array
          items:
            type: string
          minItems: 4
          maxItems: 4
          example: ["パーティーに誘うため", "仕事の相談のため", "誕生日を伝えるため", "引っ越しの手伝いを頼むため"]
        correct_answer:
          type: string
          example: "パーティーに誘うため"
        explanation:
          type: string
          example: "The passage states Tanaka-san called to invite the person to a party."
    ReadingPassage:
      type: object
      required: [id, jlpt_level, difficulty_score, word_count, source]
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          nullable: true
          example: "田中さんのパーティー"
        jlpt_level:
          type: string
          enum: [N5, N4, N3, N2, N1]
          example: "N4"
        difficulty_score:
          type: number
          format: float
          description: Raw jReadability score. Lower = harder (N1), higher = easier (N5).
          example: 6.10
        word_count:
          type: integer
          example: 110
        source:
          type: string
          enum: [jaquad, jsquad, aozora, llm_qwen3.5]
          example: "llm_qwen3.5"
        passage_text:
          type: string
          nullable: true
          description: >
            Raw Japanese text, no furigana markup. Fetched from R2 at request
            time via content_key — never stored directly in Neon. Only
            included on GET /reading/passages/{id}; omitted from list
            responses to avoid an R2 fetch per row.
          example: "来週の土曜日、田中さんから電話がありました。"
        furigana_segments:
          type: array
          nullable: true
          description: >
            Populated only for source=aozora. Null for all other sources —
            frontend falls back to client-side kuroshiro.js, same as
            sentences. Only included on GET /reading/passages/{id}.
          items:
            $ref: '#/components/schemas/FuriganaSegment'
        english_translation:
          type: string
          nullable: true
          description: Only included on GET /reading/passages/{id}.
          example: "Last week I received a phone call from Tanaka-san."
        questions:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ReadingQuestion'
          description: >
            All questions for this passage, ordered by question_order. Only
            included on GET /reading/passages/{id}. Not included in list
            responses.
    PaginatedReadingPassage:
      type: object
      required: [data, total, page, page_size]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReadingPassage'
        total:
          type: integer
          example: 200
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20


# ─────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────
paths:
...
  /kana:
    get:
      summary: List all kana characters
      tags: [kana]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: type
          in: query
          schema:
            type: string
            enum: [hiragana, katakana]
          description: Filter by hiragana or katakana
        - name: category
          in: query
          schema:
            type: string
            enum: [base, dakuten, handakuten, yoon]
          description: Filter by kana category
      responses:
        '200':
          description: Paginated list of kana
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedKana'
...
  /kanji:
    get:
      summary: List kanji characters
      description: Public endpoint. No auth required. Does not include components or sentences.
      tags: [kanji]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/JlptParam'
        - $ref: '#/components/parameters/JlptMaxParam'
        - name: grade
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 8
          description: Filter by Japanese school grade
        - name: stroke_count
          in: query
          schema:
            type: integer
          description: Filter by number of strokes
      responses:
        '200':
          description: Paginated list of kanji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedKanji'
...
  /vocab:
    get:
      summary: List vocabulary words
      description: Public endpoint. No auth required.
      tags: [vocab]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/JlptParam'
        - $ref: '#/components/parameters/JlptMaxParam'
        - name: is_common
          in: query
          schema:
            type: boolean
          description: Filter to common words only (ichi1/news1/spec1 tagged)
        - name: search
          in: query
          schema:
            type: string
          description: Search by word or reading (partial match)
      responses:
        '200':
          description: Paginated list of vocabulary words
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedVocab'
...
  /sentences:
    get:
      summary: List example sentences
      description: Public endpoint. No auth required. Paginated, searchable by Japanese text.
      tags: [sentences]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: search
          in: query
          schema:
            type: string
          description: Search within Japanese text (partial match)
      responses:
        '200':
          description: Paginated list of sentences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSentence'
...
  /components:
    get:
      summary: List all kanji components (radicals)
      description: Public endpoint. No auth required.
      tags: [components]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated list of components
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedComponent'
...
  /reading/passages:
    get:
      summary: List reading comprehension passages
      description: >
        Public endpoint. No auth required. Filterable by JLPT level.
        passage_text/furigana_segments/english_translation/questions are
        omitted here — fetch GET /reading/passages/{id} for full content.
      tags: [reading]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: jlpt_level
          in: query
          schema:
            type: string
            enum: [N5, N4, N3, N2, N1]
      responses:
        '200':
          description: Paginated list of passages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReadingPassage'
  /reading/passages/{id}:
    get:
      summary: Get a single passage with its questions
      description: >
        Public endpoint. No auth required. Fetches passage_text (and
        furigana_segments/english_translation) from R2, and all questions
        for the passage ordered by question_order.
      tags: [reading]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Passage detail including nested questions array
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingPassage'
        '404':
          description: Passage not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # ── Files (R2 uploads) ────────────────────────────────────
...